pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "uaag2"
version = "0.0.1"
description = "protein diffusion"
authors = [
  { name = "Han Tang", email = "your@email.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "biopython>=1.86",
    "fastapi==0.128.0",
    "huggingface-hub[cli]>=1.3.1",
    "lmdb>=1.7.5",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.0",
    "omegaconf>=2.3.0",
    "openbabel-wheel>=3.1.1.22",
    "pandas>=2.3.3",
    "posebusters>=0.6.3",
    "py3dmol>=2.5.3",
    "pytorch-lightning==2.6.0",
    "rdkit>=2025.9.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb==0.24.0",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = [
    "coverage==7.13.1",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.14.13",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
    "ipython>=9.9.0",
    "huggingface-hub[cli]>=1.3.1",
    "google-api-python-client>=2.188.0",
    "httpx>=0.27.0",
    "locust>=2.32.5",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv]
no-build-isolation-package = ["torch-cluster", "torch-scatter", "torch-sparse", "torch-spline-conv"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:lightning_fabric.*",
    "ignore::DeprecationWarning:pytorch_lightning.*",
    "ignore::DeprecationWarning:torch_geometric.*",
    "ignore::UserWarning:openbabel.*",
    "ignore::UserWarning:lightning_fabric.*",
    "ignore::DeprecationWarning:importlib.*",
]

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = [
    "*/_remote_module_non_scriptable.py",
]
